src/main/java/de/hbrs/team89/se1_starter_repo/ParkhausTest.java
import static org.junit.jupiter.api.Assertions.*;

class ParkhausTest {
    Parkhaus p;

    @org.junit.jupiter.api.BeforeEach
    void setUp() {
        p = new Parkhaus();
    }

    @org.junit.jupiter.api.Test
    void einUndAusfahren() {
        // Drei fahren ein (besetzt = 3)
        p.einfahren();
        p.einfahren();
        p.einfahren();

        assertEquals(3, p.getInsgesamt());
        assertEquals(3, p.getBesetzt());

        // Einer faehrt raus (besetzt = 2)
        p.ausfahren(0);

        assertEquals(2, p.getBesetzt());
        assertEquals(3, p.getInsgesamt());

        // Zwei fahren raus (besetzt = 0)
        p.ausfahren(1);
        p.ausfahren(2);

        assertEquals(0, p.getBesetzt());
        assertEquals(3, p.getInsgesamt());
    }
}